refactor: manager 패키지 삭제
- 삭제
	- ManagerPostController.java
	- ManagerNoticeController.java
	- ManagerPostService.java
	- ManagerNoticeService.java
- ManagerNotice 관련 내용을 NoticeController, NoticeService에 이전.

- #38
