+++ src/main/java/sesac/server/feed/controller/NoticeController.java
@@ -1,14 +1,23 @@
 package sesac.server.feed.controller;
 
+import static sesac.server.feed.exception.PostErrorCode.INVALID_CONTENT_SIZE;
+import static sesac.server.feed.exception.PostErrorCode.INVALID_TITLE_SIZE;
+import static sesac.server.feed.exception.PostErrorCode.REQUIRED_CONTENT;
+import static sesac.server.feed.exception.PostErrorCode.REQUIRED_NOTICE_TYPE;
+import static sesac.server.feed.exception.PostErrorCode.REQUIRED_TITLE;
+
 import jakarta.validation.Valid;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.log4j.Log4j2;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PatchMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -19,12 +28,18 @@ import org.springframework.web.bind.annotation.RestController;
 import sesac.server.auth.dto.AuthPrincipal;
 import sesac.server.auth.dto.CustomPrincipal;
 import sesac.server.common.exception.BindingResultHandler;
+import sesac.server.feed.dto.request.CreateNoticeRequest;
+import sesac.server.feed.dto.request.NoticeListRequest;
 import sesac.server.feed.dto.request.ReplyRequest;
+import sesac.server.feed.dto.request.UpdateNoticeRequest;
+import sesac.server.feed.dto.response.NoticeListResponse;
+import sesac.server.feed.dto.response.NoticeResponse;
 import sesac.server.feed.dto.response.ReplyResponse;
 import sesac.server.feed.entity.ArticleType;
 import sesac.server.feed.entity.FeedType;
 import sesac.server.feed.exception.ReplyErrorCode;
 import sesac.server.feed.service.LikesService;
+import sesac.server.feed.service.NoticeService;
 import sesac.server.feed.service.ReplyService;
 
 @Log4j2
@@ -33,14 +48,22 @@ import sesac.server.feed.service.ReplyService;
 @RequestMapping("notices/{feedType}")             // feedType - ALL, GROUP
 public class NoticeController {
 
+    private final NoticeService noticeService;
     private final LikesService likesService;
     private final ReplyService replyService;
     private final BindingResultHandler bindingResultHandler;
 
     // -----------------------------------------------------------공지 CRUD
     @GetMapping
-    public ResponseEntity<Void> getNoticeList(@PathVariable String feedType) {
-        return null;
+    public ResponseEntity<Page<NoticeListResponse>> getNoticeList(
+            @PathVariable String feedType,
+            Pageable pageable,
+            @ModelAttribute NoticeListRequest request
+    ) {
+        Page<NoticeListResponse> response = noticeService.getNoticeList(pageable, request,
+                request.type());
+
+        return ResponseEntity.ok(response);
     }
 
     @GetMapping("important")
@@ -49,35 +72,65 @@ public class NoticeController {
     }
 
     @GetMapping("{noticeId}")
-    public ResponseEntity<Void> getNotice(
-            @PathVariable String feedType, @PathVariable String noticeId
+    public ResponseEntity<NoticeResponse> getNotice(
+            @PathVariable String feedType, @PathVariable Long noticeId
     ) {
-        return null;
+        NoticeResponse response = noticeService.getNotice(noticeId);
+
+        return ResponseEntity.ok(response);
     }
 
     // 매니저 권한: 공지 생성
     @PreAuthorize("hasRole('MANAGER')")
     @PostMapping
-    public ResponseEntity<Void> createNotice(@PathVariable String feedType) {
-        return null;
+    public ResponseEntity<Void> createNotice(
+            @PathVariable String feedType,
+            @AuthPrincipal CustomPrincipal principal,
+            @Valid @RequestBody CreateNoticeRequest request,
+            BindingResult bindingResult
+    ) {
+        bindingResultHandler.handleBindingResult(bindingResult, List.of(
+                REQUIRED_TITLE,
+                INVALID_TITLE_SIZE,
+                REQUIRED_CONTENT,
+                INVALID_CONTENT_SIZE,
+                REQUIRED_NOTICE_TYPE
+        ));
+
+        noticeService.createNotice(principal.id(), request);
+
+        return ResponseEntity.ok().build();
     }
 
     // 매니저 권한: 공지 수정
     @PreAuthorize("hasRole('MANAGER')")
     @PutMapping("{noticeId}")
     public ResponseEntity<Void> updateNotice(
-            @PathVariable String feedType, @PathVariable String noticeId
+            @PathVariable String feedType,
+            @PathVariable Long noticeId,
+            @Valid @RequestBody UpdateNoticeRequest request,
+            BindingResult bindingResult
     ) {
-        return null;
+        bindingResultHandler.handleBindingResult(bindingResult, List.of(
+                INVALID_TITLE_SIZE,
+                INVALID_CONTENT_SIZE
+        ));
+
+        noticeService.updateNotice(noticeId, request);
+
+        return ResponseEntity.noContent().build();
     }
 
     // 매니저 권한: 공지 삭제
     @PreAuthorize("hasRole('MANAGER')")
     @DeleteMapping("{noticeId}")
     public ResponseEntity<Void> deleteNotice(
-            @PathVariable String feedType, @PathVariable String noticeId
+            @PathVariable String feedType,
+            @PathVariable Long noticeId
     ) {
-        return null;
+        noticeService.deleteNotice(noticeId);
+
+        return ResponseEntity.noContent().build();
     }
 
     // -----------------------------------------------------------좋아요
